backend-on-python/pyproject.toml
[tool.poetry]
authors = ["admin <admin@example.com>"]
description = "FastAPI project generated using minimal-fastapi-postgres-template."
name = "app"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

alembic = "^1.16.1"
asyncpg = "^0.30.0"
bcrypt = "^4.3.0"
fastapi = "^0.115.12"
pydantic = { extras = ["dotenv", "email"], version = "^2.11.5" }
pydantic-settings = "^2.9.1"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.41" }
python-jose = "^3.5.0"
openai = "^1.98.0"
faker = "^37.5.3"

[tool.poetry.group.dev.dependencies]
coverage = "^7.8.2"
freezegun = "^1.5.2"
greenlet = "^3.2.2"
httpx = "^0.28.1"
mypy = "^1.16.0"
pre-commit = "^4.2.0"
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.1.1"
pytest-xdist = "^3.7.0"
ruff = "^0.11.12"
uvicorn = { extras = ["standard"], version = "^0.34.3" }

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
addopts = "-vv -n auto --cov --cov-report xml --cov-report term-missing --cov-fail-under=100"
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.coverage.run]
concurrency = ["greenlet"]
omit = ["app/tests/*"]
source = ["app"]

[tool.mypy]
python_version = "3.13"
strict = true

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
# pycodestyle, pyflakes, isort, pylint, pyupgrade
ignore = ["E501"]
select = ["E", "F", "I", "PL", "UP", "W"]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["PLR0915"]
